replace update() with new conventions
KC and I decided to use a universal ducktype
method of `update_state()` which takes a timedelta,
optionally uses it, but the overall purpose is to...
update the "state," the attributes, the data of the
thing/object/sprite. Will especially be useful for
slowing down/speeding up various things.

There is no use, then, for `update()`.